Refactoring components and imports updated
