+++ src/components/Book/styles/BookCardStyle.js
