+++ src/components/Shelf/ShelfNote.js
