+++ src/components/library/Shelf.js
@@ -1,12 +1,12 @@
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { setCurrentShelf } from '../../store/actions';
-import Header from '../common/Header';
+import Header from '../Navigation/Header';
 import SearchForm from '../search/SearchForm';
-import Breadcrumbs from '../common/Breadcrumbs';
-import ShelfNote from '../common/ShelfNote';
-import BookCardList from '../common/BookCardList';
-import MyShelves from '../common/MyShelves';
+import Breadcrumbs from '../Navigation/Breadcrumbs';
+// import ShelfNote from '../Shelf/ShelfNote';
+import BookCardList from '../Book/BookCardList';
+import MyShelves from '../Shelf/MyShelves';
 import useDocumentTitle from '../../utils/hooks/useDocumentTitle';
 import styled from 'styled-components';
 import { PageView, Event } from '../../utils/tracking';
