+++ src/components/search/Search.js
@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
-import Header from '../common/Header';
-import Breadcrumbs from '../common/Breadcrumbs';
+import Header from '../Navigation/Header';
+import Breadcrumbs from '../Navigation/Breadcrumbs';
 import SearchForm from './SearchForm';
-import ShelfNote from '../common/ShelfNote';
-import BookCardList from '../common/BookCardList';
-import MyShelves from '../common/MyShelves';
+import ShelfNote from '../Shelf/ShelfNote';
+import BookCardList from '../Book/BookCardList';
+import MyShelves from '../Shelf/MyShelves';
 import useDocumentTitle from '../../utils/hooks/useDocumentTitle';
-import Loader from '../common/Loader';
+import Loader from '../Navigation/Header';
 import styled from 'styled-components';
 import { BackTop } from 'antd';
 import { PageView, Event } from '../../utils/tracking';
